A significantly more rigorous crawler.
This includes a separation between the link discoverer and the actual downloader piece (much as I previously separated the filename fixer and the crawler), which in turns involves using an actual friggin' queue to store files that will be processed (in other words, the crawler now uses an actual frontier instead of a weird stack-like thing).

This also adds proper (and thorough) logging using MonoLogger, removes largely broken APC functionality, and starts using Redis to store 4/500-files that shouldn't be reattempted on future crawls.

Beyond this, the crawler will now correctly note files that effectively redirect to themselves (most commonly because of HTTP-to-HTTPs redirects that naturally aren't recorded in the filesystem path), effectively addressing a fairly major bug of Heritrix MirrorWriter. Finally, it fleshes out the previous redirect handling in a much more ellegant way (now using Guzzle instead of the weird custom redirect handling from before), which ensures that we don't download the final resource if we already have it (but do create redirect files for every redirect along the way). Again, this is a much better behaver than Heritrix MirrorWriter itself, to the point where I would argue this crawler is actually superior to MirrorWriter for small crawls (for big ones, you still can't beat Heritrix's proper frontier).